+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changes
 0.2 (unreleased)
 ----------------
 
-- No changes yet.
+- ``node.ext.yaml.YamlNode`` and ``node.ext.yaml.YamlFile`` not provides a
+  default child factory any more.
+  [rnix]
 
 
 0.1 (2021-11-22)